requirements.txt
fastapi
uvicorn
rembg
Pillow
python-multipart
onnxruntime